requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.8.2
uvloop; platform_system!="Windows"
